e2e: common way to fetch poolName
On OpenShift we're looking for the MCP associated with the
nodes affected by the profile.

On HyperShift we're looking for the nodePool associated the
nodes affected by the profile.

This commit introduce a function to get the pool name
which is essentially MCP/NodePool name depending on
the platform detected at runtime.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>